Bug 16324: Move item search into header
This patch makes the Search button in the header into a dropdown that holds the Advanced search and the Item search.
I have rewritten this patch to be independent of Bug 16323 since that patch is so controversial.

To test:
1) Confirm that Advanced search and Item search are both under the 'Search' dropdown menu in the header and both links work as expected

Sponsored-by: Catalyst IT
Signed-off-by: Claire Gravely <c.gravely@arts.ac.uk>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Note that the CAN_user_catalogue test is useless. If a librarian does
not have it, (s)he cannot access to the staff interface.

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>