Bug 16906 - Add DataTables pagination and filter to top of saved SQL reports table
This patch adds a DataTables toolbar to the top of the saved SQL reports
table so that pagination and filtering controls are at the top as well
as the bottom.

To test, apply the patch and go to Reports -> Saved reports. There
should be controls at the top of the table of saved reports for setting
number of results, pagination, and searching.

Signed-off-by: Marc Véron <veron@veron.ch>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>